Virtualhospital tanya

+++ src/Screens/Components/VirtualHospitalComponents/TaskSectionVH/index.js
@@ -348,10 +348,45 @@ class Index extends Component {
                 : "react-confirm-alert-body"
             }
           >
-            <h1>Remove the Comment ?</h1>
+            <h1 >Remove the Comment ?</h1>
             <p>Are you sure to remove this Comment?</p>
             <div className="react-confirm-alert-button-group">
               <button onClick={onClose}>No</button>
+              
+              <button
+                onClick={() => {
+                  this.removebtn(index);
+                }}
+              >
+                Yes
+              </button>
+              
+            </div>
+          </div>
+        );
+      },
+    });
+  };
+  removebtn = (index) => {
+    this.setState({ message: null, openTask: false });
+    confirmAlert({
+      customUI: ({ onClose }) => {
+        return (
+          <div
+            className={
+              this.props.settings &&
+                this.props.settings.setting &&
+                this.props.settings.setting.mode &&
+                this.props.settings.setting.mode === "dark"
+                ? "dark-confirm react-confirm-alert-body"
+                : "react-confirm-alert-body"
+            }
+          >
+            <h1 class ="alert-btn">Remove Comment ?</h1>
+            <p>Are you really want to remove this Comment?</p>
+            <div className="react-confirm-alert-button-group">
+              <button onClick={onClose}>No</button>
+              
               <button
                 onClick={() => {
                   this.deleteClickComment(index);
@@ -360,6 +395,7 @@ class Index extends Component {
               >
                 Yes
               </button>
+            
             </div>
           </div>
         );
@@ -525,6 +561,40 @@ class Index extends Component {
           >
             <h1>Remove the Task ?</h1>
             <p>Are you sure to remove this Task?</p>
+            <div className="react-confirm-alert-button-group">
+              <button onClick={onClose}>No</button>
+              <button
+                onClick={() => {
+                  this.removeTask2(id);
+                  // onClose();
+                }}
+              >
+                Yes
+              </button>
+            </div>
+          </div>
+        );
+      },
+    });
+  };
+
+  removeTask2 = (id) => {
+    this.setState({ message: null, openTask: false });
+    confirmAlert({
+      customUI: ({ onClose }) => {
+        return (
+          <div
+            className={
+              this.props.settings &&
+                this.props.settings.setting &&
+                this.props.settings.setting.mode &&
+                this.props.settings.setting.mode === "dark"
+                ? "dark-confirm react-confirm-alert-body"
+                : "react-confirm-alert-body"
+            }
+          >
+            <h1 class="alert-btn">Remove Task?</h1>
+            <p>Are you really want to remove this Task?</p>
             <div className="react-confirm-alert-button-group">
               <button onClick={onClose}>No</button>
               <button
@@ -1161,6 +1231,7 @@ class Index extends Component {
                                         name="comment"
                                           onChange={(e) => this.oNEditText(e, index) 
                                           }
+
                                           onKeyDown={this.onKeyUp}
                                           value={data?.comment}
                                         >
